refactor: extract MicrophonePermissionStatus.swift from PermissionAndPreflightTypes.swift
Splits the `MicrophonePermissionStatus` raw-string enum out of
PermissionAndPreflightTypes.swift.

Byte-identical move of lines 10-18. Both files import Foundation.

PermissionAndPreflightTypes.swift: 83 → 73 lines. Tests: 667.
