+++ Sources/BugNarrator/Services/MicrophonePermissionStatus.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+enum MicrophonePermissionStatus: String, Equatable {
+    case notDetermined
+    case granted
+    case denied
+    case restricted
+    case unavailable
+    case captureSetupFailed
+    case unknownError
+}
